[10193] Use for Transport::Update arg uint32 type as for other world objects

+++ src/game/MapManager.cpp
@@ -250,7 +250,7 @@ MapManager::Update(uint32 diff)
         iter->second->Update((uint32)i_timer.GetCurrent());
 
     for (TransportSet::iterator iter = m_Transports.begin(); iter != m_Transports.end(); ++iter)
-        (*iter)->Update(i_timer.GetCurrent());
+        (*iter)->Update((uint32)i_timer.GetCurrent());
 
     i_timer.SetCurrent(0);
 }
